refactor: :art: Remove additional data file
Related to #60
